requirements.txt
openai
pydantic
python-dotenv 
pandas
numpy
bs4
selenium
# Dev packages
mypy==0.982
black==22.10.0
isort==5.10.1
pytest==7.2.1
greenlet
